move menu extension types from registry to menu module
