+++ src/packages/core/extension-registry/extensions/index.ts
@@ -7,10 +7,6 @@ export type * from './external-login-provider-element.interface.js';
 export type * from './external-login-provider.extension.js';
 export type * from './global-context.extension.js';
 export type * from './header-app.extension.js';
-export type * from './menu-item-element.interface.js';
-export type * from './menu-item-element.interface.js';
-export type * from './menu-item.extension.js';
-export type * from './menu.extension.js';
 export type * from './mfa-login-provider.extension.js';
 export type * from './preview-app.extension.js';
 export type * from './repository.extension.js';
