+++ src/packages/core/menu/index.ts
@@ -2,5 +2,8 @@ export * from './components/index.js';
 export * from './menu-tree-structure-workspace-context-base.js';
 export * from './menu-variant-tree-structure-workspace-context-base.js';
 export * from './types.js';
+export type * from './menu-item-element.interface.js';
+export type * from './menu-item.extension.js';
+export type * from './menu.extension.js';
 
 export type { UmbMenuStructureWorkspaceContext } from './menu-structure-workspace-context.interface.js';
